completed
A House Price Prediction Model in machine learning is a predictive model that estimates the market price of a property (house) based on various input features or attributes of the property. These features typically include both structural and environmental factors that could affect a property's value, such as its size, location, number of bedrooms, number of bathrooms, age of the house, proximity to amenities, neighborhood crime rate, and other socio-economic indicators.